[Mono.Android.GoogleMaps] Generate merged API assemblies.
(More of a "this is how things _could_ be done" rather than
"how things _should_ be done", as the scenario outlined below isn't
actually the case with the Google Maps API.)

Imagine you have a Java library which changes over time, like, oh,
android.jar. ;-)

The problem: using `mcw-gen build` will result in an assembly which
matches the source .jar 1:1 (module some "niceness" changes -- added
events, etc.). This means that if a member _moves_ between .jar
versions, you can't necessarily use an assembly built on version N+1
on a platform that only provides version N ("the MotionEvent issue").

Which brings us to `mcw-gen merge`, new to Mono for Android 4.1.1.
This allows "merging" API descriptions so that instead of having an
API description which is a 1:1 representation of the underlying .jar,
it is instead the union of the types and mebers across versions. If a
method is moved into an introduced base class, the merged API will
contain both the introduced base class + method AND the "original"
member.

This allows an assembly binding N+1 to reference the "original" and
execute sanely on a platform that has version N (as long as you don't
access members that only exist in version N+1).

Generating such a merged assembly requires 4 steps:

 1. Generate the 1:1 API description for all desired versions, e.g.
    maps.jar v8 requires descriptions for v4, v7, and v8.

 2. Merge these API descriptions into a single description:

	mcw-gen merge -o merged.xml source1.xml source2.xml ...

 3. Generate sources based on the merged XML

 	mcw-gen bind -i merged.xml -o dir ..

 4. Compile the sources:

	smcs -t:library -out:binding.dll dir/src/generated/*.cs ...