[GoogleMaps] Rework maps.jar resolution.
The problem: Installing the API 16 Google Addons SDK causes the build
to break (!). Why? Because the e.g. API 4 adds path is:

	add-ons/addon-google_apis-google_inc_-4

while the API 16 path is:

	add-ons/addon-google_apis-google-16

No problem, right? Wrong:

	JARS      = $(shell find $(ANDROID_SDK_PATH)/ -name maps.jar)
	ADDONDIRS = $(patsubst %/libs/maps.jar,%,$(JARS))
	ADDONDIR  = $(shell echo $(firstword $(ADDONDIRS)) | sed -e 's/^\(.*-\).*$$/\1/')

$(JARS) contains every maps.jar within the Android SDK, no problem.
$(ADDONDIRS) contains all add-ons/addon-* directories which eventually
contain a maps.jar file; no problem.

The problem is $(ADDONDIR), which assumes that every such
add-ons/addon-* directory has a common prefix, with the only
difference being the API number. This is NO LONGER the case. Due to
path lookup ordering, $(ADDONDIRS) contained:

	add-ons/addon-google_apis-google-16
	add-ons/addon-google_apis-google_inc_-4

i.e. API 16 was first, so when we tried to generate the API 4
description, we tried to find add-ons/addon-google_apis-google-4,
which doesn't exist.

Doh!

The solution? Fuck that shit; `find` FOUND a maps.jar to use. Don't
look for a common directory name prefix, just use the FOUND NAME. This
does mean that we can't use $(JARS) as a dependency (we're only going
to find what's installed), so we use an explicit file check instead.