Obsolete ALL THE ASSEMBLIES!
Mwa-ha-ha-ha-ha *cough* *splutter*

We've shipped Mono.Android.GoogleMaps.dll,
Mono.Android.Support.v4.dll, and Mono.Android.Support.v13.dll since
close to the launch of Mono for Android.

There's just one current problem with them: they're old.

Mono.Android.GoogleMaps.dll bound the maps.jar addon, which is no
longer distributed anymore, and it's no longer possible to obtain
certificates to use the library. It's dead. It's replacement is the
Google Play Services library, and the corresponding Component:

	http://components.xamarin.com/view/googleplayservices

Mono.Android.Support*.dll are in a slightly different boat: They at
least still exist in a (mostly) similar form. The problem is that
the underlying Java library appeared to break backward
compatibility [0] awhile ago, preventing us from easily updating the
assembly, and we now have the Component to take their places, which
are considerably more up to date and have gotten FAR more work:

	http://components.xamarin.com/view/xamandroidsupportv4-18
	http://components.xamarin.com/view/xamandroidsupportv13-18

TL;DR: We're obsoleting these assemblies so that we can remove them in
a future Xamarin.Android release, and direct people to use the newer,
and considerably better, alternatives.

[0]: This is on memory, but at the time the support library contained
types with platform names, like JellyBeanSomethingOrOther, and these
had "disappeared" after the update. In retrospect, it seems that
Google changed the support lib from being "just" a .jar to being a
core .jar + Source code, and many of the missing types were now
available as source.

This was in a cruch time, so the "retrospect" didn't occur until
months later, by which time the Xamarin Component replacement existed
and was using a newer version of the library.

We have enough maintenance work with the core; we more than happy to
have the Components team take over these bindings.